requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
mcrcon==0.7.0
mysql-connector-python==8.0.33